feat(Brandfetch Node): use explicit route types and simplify operations
- Add Type selector (domain, ticker, crypto, isin) and free-form Identifier
  input, using Brandfetch's explicit /v2/brands/{type}/{identifier} routes
  to avoid identifier collisions.
- Replace per-field operations (color, company, font, industry, logo) with
  two operations: Brand Data (full payload) and Brand Logos.
- Remove the binary download flow and the unused fetchAndPrepareBinaryData
  helper.
- Update unit tests to cover all four explicit route types.

feat(nodes): add versioning and colors operation to Brandfetch node
