+++ packages/nodes-base/nodes/Brandfetch/v1/GenericFunctions.ts
